Move CRM Web API methods back into the CRM Service
